refactor: remove stores barrel index files and use direct imports [WTEL-9717](https://webitel.atlassian.net/browse/WTEL-9717)
