+++ src/modules/configuration/modules/lookups/modules/contact-groups/modules/add-contacts-in-group/components/add-contacts-in-group-filters-panel.vue
@@ -14,7 +14,7 @@
 import { TableFiltersPanelComponent as TableFiltersPanel } from '@webitel/ui-datalist/filters';
 
 import { filtersOptions } from '../configs/filtersOptions';
-import { createAddContactsInGroupDatalistStore } from '../stores';
+import { createAddContactsInGroupDatalistStore } from '../stores/datalist/addContactsInGroupDatalistStore';
 
 const props = defineProps<{
 	useTableStore: ReturnType<typeof createAddContactsInGroupDatalistStore>;
