+++ src/modules/configuration/modules/lookups/modules/contact-groups/modules/add-contacts-in-group/components/add-contacts-in-group-popup.vue
@@ -107,7 +107,7 @@ import { useTableEmpty } from '@webitel/ui-sdk/src/modules/TableComponentModule/
 import { computed, ref } from 'vue';
 import { useI18n } from 'vue-i18n';
 
-import { createAddContactsInGroupDatalistStore } from '../stores';
+import { createAddContactsInGroupDatalistStore } from '../stores/datalist/addContactsInGroupDatalistStore';
 import AddContactInGroupSearchBar from './add-contact-in-group-search-bar.vue';
 import AddContactsInGroupFiltersPanel from './add-contacts-in-group-filters-panel.vue';
