--- src/modules/configuration/modules/lookups/modules/contact-groups/modules/add-contacts-in-group/stores/index.ts
@@ -1,1 +1,0 @@
-export * from './datalist/addContactsInGroupDatalistStore';
