--- src/modules/configuration/modules/lookups/modules/contact-groups/modules/conditions/stores/index.ts
@@ -1,2 +1,0 @@
-export * from './card/contactGroupConditionsCardStore';
-export * from './datalist/contactGroupConditionsDatalistStore';
