--- src/modules/configuration/modules/lookups/modules/contact-groups/modules/contacts/stores/index.ts
@@ -1,1 +1,0 @@
-export * from './datalist/contactGroupContactsDatalistStore';
