--- src/modules/configuration/modules/lookups/modules/contact-groups/stores/index.ts
@@ -1,3 +1,0 @@
-export * from './card/contactGroupsCardStore';
-export * from './datalist/contactGroupsDatalistStore';
-export * from './permissions/contactGroupsPermissionsStore';
